// ClientApp/src/app/basic-setup/models/BookTitleInfo.ts
export interface BookTitleInfo {
    bookTitleInfoId: number;
    bookTitleName: string;
    baseSchoolNameId:number;
    titleBangla: string;
    bookSubtitle: string;
    remarks: string;
    createDate: Date;
    createUserId: string;
    updateDate: Date;
    updateUserId: string;
    isActive: boolean;
}